# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Zinc Search engine API documents https://docs.zincsearch.com",
        "title": "Zinc Search engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Zinc Search",
            "url": "https://www.zincsearch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4080",
    "basePath": "/",
    "paths": {
...
        "/api/_bulk": {
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "Bulk documents",
                "operationId": "Bulk",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseRecordCount"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/{index}/_doc/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "Create or update document with id",
                "operationId": "IndexDocumentWithID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document",
                        "name": "document",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "Delete document",
                "operationId": "DeleteDocument",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseDocument"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/es/_bulk": {
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "ES bulk documents",
                "operationId": "ESBulk",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "meta.HTTPResponseDocument": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "meta.HTTPResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "meta.HTTPResponseID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseRecordCount": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "record_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}